+++ viajeros/src/components/ViajeroLoader/ViajeroLoader.tsx
@@ -1,9 +1,10 @@
+import { VIAJERO_GREEN } from "@/consts"
 import { Center, Loader } from "@mantine/core"
 
 export const ViajeroLoader = () => {
   return (
     <Center h="100vh">
-      <Loader color="#65a773" />
+      <Loader color={VIAJERO_GREEN} />
     </Center>
   )
 }
